refactor: into nuxt server side rendering
