+++ public/components/SchoolList.vue
