--- public/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="x-ua-compatible" content="ie=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>BentoBox</title>
-    <link rel="stylesheet" href="//fonts.googleapis.com/css?family=Roboto:300,400,500,700,400italic">
-    <link href="//fonts.googleapis.com/css?family=Dosis:200|Rubik:900" rel="stylesheet">
-    <link rel="stylesheet" href="//fonts.googleapis.com/icon?family=Material+Icons">
-  </head>
-  <body>
-    <div id="app">
-      <router-view></router-view>
-    </div>
-    <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyCNsW5TODUbmv2K_0pswqz_P09-e6YeY8s">
-    </script>
-    <!-- built files will be auto injected -->
-  </body>
-</html>
