+++ public/package.json
@@ -1,12 +1,13 @@
 {
-  "name": "school-donator",
+  "name": "BentoBox",
   "version": "1.0.0",
-  "description": "A service to simplify school donation process",
+  "description": "BentoBox helps to pay unpaid lunch balance for school kids",
   "author": "Eric Liao <rcliao01@gmail.com>",
   "private": true,
   "scripts": {
-    "dev": "node build/dev-server.js",
-    "build": "node build/build.js",
+    "dev": "nuxt",
+    "build": "nuxt build",
+    "generate": "nuxt generate",
     "unit": "karma start test/unit/karma.conf.js --single-run",
     "e2e": "node test/e2e/runner.js",
     "test": "npm run unit && npm run e2e",
@@ -14,34 +15,19 @@
   },
   "dependencies": {
     "axios": "^0.15.3",
-    "vue": "^2.1.0",
-    "vue-material": "^0.6.0",
-    "vue-router": "^2.1.1"
+    "node-sass": "^4.3.0",
+    "nuxt": "latest",
+    "sass-loader": "^4.1.1",
+    "vue-material": "^0.6.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.4.0",
-    "babel-core": "^6.0.0",
-    "babel-eslint": "^7.0.0",
-    "babel-loader": "^6.0.0",
-    "babel-plugin-transform-runtime": "^6.0.0",
-    "babel-preset-es2015": "^6.0.0",
-    "babel-preset-stage-2": "^6.0.0",
-    "babel-register": "^6.0.0",
+    "chai": "^3.5.0",
     "chalk": "^1.1.3",
     "connect-history-api-fallback": "^1.1.0",
-    "css-loader": "^0.25.0",
     "eslint": "^3.7.1",
     "eslint-friendly-formatter": "^2.0.5",
     "eslint-loader": "^1.5.0",
     "eslint-plugin-html": "^1.3.0",
-    "eventsource-polyfill": "^0.9.6",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "function-bind": "^1.0.2",
-    "html-webpack-plugin": "^2.8.1",
-    "http-proxy-middleware": "^0.17.2",
-    "json-loader": "^0.5.4",
     "karma": "^1.3.0",
     "karma-coverage": "^1.1.1",
     "karma-mocha": "^1.2.0",
@@ -52,28 +38,9 @@
     "karma-webpack": "^1.7.0",
     "lolex": "^1.4.0",
     "mocha": "^3.1.0",
-    "chai": "^3.5.0",
+    "postcss-cssnext": "^2.9.0",
     "sinon": "^1.17.3",
-    "sinon-chai": "^2.8.0",
-    "inject-loader": "^2.0.1",
-    "isparta-loader": "^2.0.0",
-    "phantomjs-prebuilt": "^2.1.3",
-    "chromedriver": "^2.21.2",
-    "cross-spawn": "^4.0.2",
-    "nightwatch": "^0.9.8",
-    "selenium-server": "2.53.1",
-    "semver": "^5.3.0",
-    "opn": "^4.0.2",
-    "ora": "^0.3.0",
-    "shelljs": "^0.7.4",
-    "url-loader": "^0.5.7",
-    "vue-loader": "^10.0.0",
-    "vue-style-loader": "^1.0.0",
-    "vue-template-compiler": "^2.1.0",
-    "webpack": "^1.13.2",
-    "webpack-dev-middleware": "^1.8.3",
-    "webpack-hot-middleware": "^2.12.2",
-    "webpack-merge": "^0.14.1"
+    "sinon-chai": "^2.8.0"
   },
   "engines": {
     "node": ">= 4.0.0",
