--- public/src/assets/app.css
@@ -1,50 +1,0 @@
-body, #app, main {
-  height: 100%;
-}
-
-h1 {
-  font-family: 'Rubik', sans-serif;
-}
-
-h2 {
-  font-family: 'Dosis', sans-serif;
-}
-
-.md-card .md-card-header {
-  padding: 1.5em;
-  padding-bottom: 0;
-}
-.md-card .md-card-content {
-  font-size: 16px;
-  padding: 1.5em;
-  line-height: 1.5em;
-  line-space: 1.2em;
-}
-
-.md-whiteframe-2dp {
-  margin-bottom: 1em;
-}
-
-.container {
-  width: 60em;
-  margin: 0 auto;
-}
-
-@media (max-width: 960px) {
-  .container {
-    width: 100%;
-  }
-}
-
-.flex {
-  display: flex;
-}
-
-.column {
-  flex-direction: column;
-}
-
-.flex.center {
-  justify-content: center;
-  align-items: center;
-}
