--- public/test/unit/.eslintrc
@@ -1,9 +1,0 @@
-{
-  "env": {
-    "mocha": true
-  },
-  "globals": {
-    "expect": true,
-    "sinon": true
-  }
-}
