--- public/test/unit/specs/Hello.spec.js
@@ -1,13 +1,0 @@
-import Vue from 'vue'
-import Hello from 'src/components/Hello'
-
-describe('Hello.vue', () => {
-  it('should render correct contents', () => {
-    const vm = new Vue({
-      el: document.createElement('div'),
-      render: (h) => h(Hello)
-    })
-    expect(vm.$el.querySelector('.hello h1').textContent)
-      .to.equal('Welcome to Your Vue.js App')
-  })
-})
